refactor(codex): move runner goal helper into package
